Add exception for V flag clear on cache name edit
If the user has already verified the SAS on the current call we don't
need to clear the V flag or the Secure state.  We're only worried here
with a Secure state that results from the V flag being set in the
cache when the cache name is wrong.  An SAS verify on the current call
has ruled out this attack.

This also solves a conundrum about UI state progression on the initial
verify.  Some clients prompt the user to edit the cache name
immediately after the SAS verify.  Clearly we don't want to clear the
V flag in that case.  This change provides the general solution.